Create Mark module
This creates most of the mark-related types, but not MarkChannel
